- cleaning and reducing the amount of files, perfer to use just handlers and middleware - change config option in json/yaml of clientid to client-id - added additional unit test for the sessions and cookie methods - added support for password for redis - added the ability to proxy to a unix socket - cleaned up the handlers and middleware into respective files, i dont like lots of files laying around
