--- handlers_entry.go
@@ -1,64 +1,0 @@
-/*
-Copyright 2015 All rights reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"strings"
-
-	"github.com/gin-gonic/gin"
-)
-
-const (
-	// cxEnforce is the tag name for a request requiring
-	cxEnforce = "Enforcing"
-)
-
-//
-// entryPointHandler checks to see if the request requires authentication
-//
-func (r oauthProxy) entryPointHandler() gin.HandlerFunc {
-	return func(cx *gin.Context) {
-		if strings.HasPrefix(cx.Request.URL.Path, oauthURL) {
-			cx.Next()
-			return
-		}
-
-		// step: check if authentication is required - gin doesn't support wildcard url, so we have have to use prefixes
-		for _, resource := range r.config.Resources {
-			if strings.HasPrefix(cx.Request.URL.Path, resource.URL) {
-				if resource.WhiteListed {
-					break
-				}
-				// step: inject the resource into the context, saves us from doing this again
-				if containedIn("ANY", resource.Methods) || containedIn(cx.Request.Method, resource.Methods) {
-					cx.Set(cxEnforce, resource)
-				}
-				break
-			}
-		}
-		// step: pass into the authentication and admission handlers
-		cx.Next()
-
-		// step: add a custom headers to the request
-		for k, v := range r.config.Header {
-			cx.Request.Header.Set(k, v)
-		}
-		// step: check the request has not been aborted and if not, proxy request
-		if !cx.IsAborted() {
-			r.proxyHandler(cx)
-		}
-	}
-}
