+++ handlers_test.go
@@ -82,80 +82,6 @@ func TestExpirationHandler(t *testing.T) {
 	}
 }
 
-func TestCrossSiteHandler(t *testing.T) {
-	kc := newFakeKeycloakProxy(t)
-	handler := kc.crossSiteHandler()
-
-	cases := []struct {
-		Cors    *CORS
-		Headers map[string]string
-	}{
-		{
-			Cors: &CORS{
-				Origins: []string{"*"},
-			},
-			Headers: map[string]string{
-				"Access-Control-Allow-Origin": "*",
-			},
-		},
-		{
-			Cors: &CORS{
-				Origins: []string{"*", "https://examples.com"},
-				Methods: []string{"GET"},
-			},
-			Headers: map[string]string{
-				"Access-Control-Allow-Origin":  "*,https://examples.com",
-				"Access-Control-Allow-Methods": "GET",
-			},
-		},
-	}
-
-	for i, c := range cases {
-		// step: get the config
-		kc.config.CORS = c.Cors
-		// call the handler and check the responses
-		context := newFakeGinContext("GET", "/oauth/test")
-		handler(context)
-		// step: check the headers
-		for k, v := range c.Headers {
-			value := context.Writer.Header().Get(k)
-			if value == "" {
-				t.Errorf("case %d, should have had the %s header set, headers: %v", i, k, context.Writer.Header())
-				continue
-			}
-			if value != v {
-				t.Errorf("case %d, expected: %s but got %s", i, k, value)
-			}
-		}
-	}
-}
-
-func TestSecurityHandler(t *testing.T) {
-	kc := newFakeKeycloakProxy(t)
-	handler := kc.securityHandler()
-	context := newFakeGinContext("GET", "/")
-	handler(context)
-	if context.Writer.Status() != http.StatusOK {
-		t.Errorf("we should have received a 200")
-	}
-
-	kc = newFakeKeycloakProxy(t)
-	kc.config.Hostnames = []string{"127.0.0.1"}
-	handler = kc.securityHandler()
-	handler(context)
-	if context.Writer.Status() != http.StatusOK {
-		t.Errorf("we should have received a 200 not %d", context.Writer.Status())
-	}
-
-	kc = newFakeKeycloakProxy(t)
-	kc.config.Hostnames = []string{"127.0.0.2"}
-	handler = kc.securityHandler()
-	handler(context)
-	if context.Writer.Status() != http.StatusInternalServerError {
-		t.Errorf("we should have received a 500 not %d", context.Writer.Status())
-	}
-}
-
 func TestHealthHandler(t *testing.T) {
 	proxy := newFakeKeycloakProxy(t)
 	context := newFakeGinContext("GET", healthURL)
